requirements.txt
fastapi
uvicorn
chromadb-client
pydantic
python-dotenv
openai